+++ colors/eldritch-glow-aqua.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-glow-aqua"
-require("eldritch").load("eldritch_glow_aqua")