+++ colors/eldritch-glow-crimson.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-glow-crimson"
-require("eldritch").load("eldritch_glow_crimson")