+++ colors/eldritch-glow-fuchsia.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-glow-fuchsia"
-require("eldritch").load("eldritch_glow_fuchsia")