+++ colors/eldritch-glow-orchid.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-glow-orchid"
-require("eldritch").load("eldritch_glow_orchid")