+++ colors/eldritch-glow-solar.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-glow-solar"
-require("eldritch").load("eldritch_glow_solar")