+++ colors/eldritch-glow-violet.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-glow-violet"
-require("eldritch").load("eldritch_glow_violet")