+++ colors/eldritch-gruvbox-azure.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-gruvbox-azure"
+require("eldritch").load("gruvbox_azure")