+++ colors/eldritch-gruvbox-crimson.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-gruvbox-crimson"
+require("eldritch").load("gruvbox_crimson")