+++ colors/eldritch-gruvbox-emerald.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-gruvbox-emerald"
+require("eldritch").load("gruvbox_emerald")