+++ colors/eldritch-kanagawa-azure.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-kanagawa-azure"
+require("eldritch").load("kanagawa_azure")