+++ colors/eldritch-kanagawa-crimson.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-kanagawa-crimson"
+require("eldritch").load("kanagawa_crimson")