+++ colors/eldritch-kanagawa-emerald.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-kanagawa-emerald"
+require("eldritch").load("kanagawa_emerald")