+++ colors/eldritch-kanagawa-fusion-glow.lua
@@ -1,2 +1,0 @@
-vim.g.colors_name = "eldritch-kanagawa-fusion-glow"
-require("eldritch").load("kanagawa_fusion_glow")