+++ colors/eldritch-rosepine-azure.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-rosepine-azure"
+require("eldritch").load("rosepine_azure")