+++ colors/eldritch-rosepine-crimson.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-rosepine-crimson"
+require("eldritch").load("rosepine_crimson")