+++ colors/eldritch-rosepine-emerald.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-rosepine-emerald"
+require("eldritch").load("rosepine_emerald")