+++ colors/eldritch-tokyonight-azure.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-tokyonight-azure"
+require("eldritch").load("tokyonight_azure")