+++ colors/eldritch-tokyonight-crimson.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-tokyonight-crimson"
+require("eldritch").load("tokyonight_crimson")