+++ colors/eldritch-tokyonight-emerald.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-tokyonight-emerald"
+require("eldritch").load("tokyonight_emerald")