+++ colors/eldritch-vscode-dark-azure.lua
@@ -1,0 +1,2 @@
+vim.g.colors_name = "eldritch-vscode-dark-azure"
+require("eldritch").load("vscode_dark_azure")